backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydicom==2.4.4
Pillow==10.2.0
numpy==1.26.3
anthropic>=0.40.0
aiofiles==23.2.1
